Added test environment and seperated .c and .h to be able to test this code
