--- CONTRIBUTING.md
@@ -1,45 +1,0 @@
-# Contributing to the project
-
-Hey there! Thanks for looking into how you're able to contribute to the project!
-Check the 'issues' tab on github to check the open issues. 
-
-_If you've any questions, feel free to reach out to <jmakhack@protonmail.com>._
-
-## Build
-
-_To build the project, execute the following steps:_
-
-```
-- make clean
-- make all
-- ./mya -v
-```
-
-## Contributing
-
-_In case you'd like to contribute to this project where are a couple things that you'll have to do: _
-
-1. Fork the project from [https://github.com/jmakhack/myanimelist-cli](https://github.com/jmakhack/myanimelist-cli)
-1. Clone the github project which you just forked
-1. Make the adjustments to solve the issue/ ticket
-1. Add your changes and push them to master
-1. Visit [https://github.com/jmakhack/myanimelist-cli](https://github.com/jmakhack/myanimelist-cli) and press on the 'Contribute' button.
-1. Create the pull request, and wait for reviewing or approval.
-
-And that's it! Thanks for contributing to the project! 
-
-## Code quality/ conventions
-_Please apply this style to your code before setting up the pull request, so we're able to remain a similar and professional code style._
-
-For coding in C, it's a good practice to follow along the following code standard: [https://users.ece.cmu.edu/~eno/coding/CCodingStandard.html](https://users.ece.cmu.edu/~eno/coding/CCodingStandard.html) 
-
-## Testing
-
-_At moment of writing, testing is not possible yet_
-
-## Included libraries
-
-_The following artifacts / tools are used by and included in the extension as-is:_
-
-- Curl: [https://curl.se/libcurl/c/](https://curl.se/libcurl/c/)
-- Json-C: [http://json-c.github.io/json-c/](http://json-c.github.io/json-c/)
